requirements.txt
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.11.0
blinker==1.9.0
chardet==6.0.0.post1
click==8.3.1
colorama==0.4.6
et_xmlfile==2.0.0
fastapi==0.121.3
Flask==3.1.2
flask-cors==6.0.2
greenlet==3.2.4
h11==0.16.0
idna==3.11
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.3
mysql-connector-python==8.0.33
numpy==2.3.5
openpyxl==3.1.5
pandas==2.3.3
protobuf==3.20.3
pydantic==2.12.4
pydantic_core==2.41.5
PyMySQL==1.1.2
pyodbc==5.3.0
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
pytz==2025.2
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.44
starlette==0.50.0
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.3
uvicorn==0.38.0
Werkzeug==3.1.4
